% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hi world fun}
\usage{
hello(x)
}
\arguments{
\item{x}{Name of somown}
}
\value{
The output form \code{\link{print}}
}
\description{
Hi world fun
}
\examples{
hello("Verena")
\dontrun{
hello("Karsten")
}
}
